refactor: Simplify trivy packages structure
