--- .github/workflows/codeql-analysis.yml
@@ -1,34 +1,0 @@
-name: CodeQL
-on:
-  push:
-    branches:
-      - master
-  pull_request:
-    branches:
-      - master
-  schedule:
-    - cron: "0 16 * * 6"
-jobs:
-  analyse:
-    name: Analyse
-    runs-on: ubuntu-18.04
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-        with:
-          # We must fetch at least the immediate parents so that if this is
-          # a pull request then we can checkout the head.
-          fetch-depth: 2
-      # If this run was triggered by a pull request event, then checkout
-      # the head of the pull request instead of the merge commit.
-      - name: Checkout head of the pull request
-        run: git checkout HEAD^2
-        if: ${{ github.event_name == 'pull_request' }}
-      - name: Initialize CodeQL
-        uses: github/codeql-action/init@v1
-        with:
-          languages: go
-      - name: Autobuild
-        uses: github/codeql-action/autobuild@v1
-      - name: Perform CodeQL analysis
-        uses: github/codeql-action/analyze@v1
