+++ Makefile
@@ -66,7 +66,8 @@ itests-starboard: check-env get-ginkgo
 	github.com/aquasecurity/starboard/pkg/kubehunter,\
 	github.com/aquasecurity/starboard/pkg/polaris,\
 	github.com/aquasecurity/starboard/pkg/polaris/crd,\
-	github.com/aquasecurity/starboard/pkg/find/vulnerabilities/trivy,\
+	github.com/aquasecurity/starboard/pkg/trivy,\
+	github.com/aquasecurity/starboard/pkg/find/vulnerabilities,\
 	github.com/aquasecurity/starboard/pkg/vulnerabilityreport \
 	./itest/starboard
 
@@ -82,7 +83,8 @@ itests-starboard-operator: check-env get-ginkgo
 	github.com/aquasecurity/starboard/pkg/operator/controller/job,\
 	github.com/aquasecurity/starboard/pkg/operator/controller/pod,\
 	github.com/aquasecurity/starboard/pkg/operator/logs,\
-	github.com/aquasecurity/starboard/pkg/operator/trivy \
+	github.com/aquasecurity/starboard/pkg/trivy,\
+	github.com/aquasecurity/starboard/pkg/vulnerabilityreport \
 	./itest/starboard-operator
 
 check-env:
