+++ pkg/find/vulnerabilities/scanner_test.go
@@ -1,0 +1,1 @@
+package vulnerabilities_test
