--- pkg/operator/trivy/scanner_test.go
@@ -1,1 +1,0 @@
-package trivy
