+++ pkg/trivy/model.go
