+++ pkg/trivy/model_test.go
