+++ pkg/trivy/scanner_test.go
